% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxiquant-package.R, R/oxiquant.R
\docType{package}
\name{oxiquant}
\alias{oxiquant}
\title{\code{oxiquant} package}
\usage{
oxiquant(msgf = FALSE, process_ms1 = FALSE, extract_ion_current = TRUE)
}
\arguments{
\item{msgf}{Run MS-GF+ MS2 search engine, logical, default = FALSE.}

\item{process_ms1}{Run mzcharge for MS1 data processing, logical, default = FALSE.}

\item{extract_ion_current}{Extract ion current for intact and oxidized peptides, logical, default = TRUE.}
}
\value{
oxiquant saves tables in Global Environment for interactive use and
writes corresponding CVS files in working directory
}
\description{
Automated analysis of mass spectrometry data collected during hydroxyl
radical footprinting experiment
}
